OpenCL: rename to follow the file naming conventions
