+++ src/opencl_cryptmd5_fmt_plug.c
@@ -57,7 +57,7 @@ static const char * warn[] = {
 };
 
 //This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 /* ------- Helper functions ------- */
