+++ src/opencl_keyring_fmt_plug.c
@@ -103,7 +103,7 @@ static const char * warn[] = {
 };
 
 //This file contains auto-tuning routine(s). It has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 /* ------- Helper functions ------- */
