+++ src/opencl_keystore_fmt_plug.c
@@ -94,7 +94,7 @@ static cl_mem mem_in, mem_out, mem_salt;
 static cl_int cl_err;
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char * warn[] = {
