+++ src/opencl_lotus5_fmt_plug.c
@@ -59,7 +59,7 @@ static cl_mem cl_tx_keys, cl_tx_binary;
 #define SEED			256
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char *warn[] = {
