+++ src/opencl_o5logon_fmt_plug.c
@@ -92,7 +92,7 @@ static unsigned int *saved_plain, *saved_idx;
 static unsigned int key_idx;
 static struct fmt_main *self;
 
-#include "opencl-autotune.h" // Must come after auto-tune definitions
+#include "opencl_autotune.h" // Must come after auto-tune definitions
 #include "memdbg.h"
 
 static size_t get_task_max_work_group_size()
