+++ src/opencl_odf_aes_fmt_plug.c
@@ -102,7 +102,7 @@ static struct fmt_tests tests[] = {
 static struct fmt_main *self;
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char * warn[] = {
