+++ src/opencl_office_fmt_plug.c
@@ -130,7 +130,7 @@ static const char * warn[] = {
 static int split_events[] = { 3, -1, -1 };
 
 //This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 /* ------- Helper functions ------- */
