+++ src/opencl_oldoffice_fmt_plug.c
@@ -116,7 +116,7 @@ static struct fmt_main *self;
 #define SEED			1024
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char *warn[] = {
