+++ src/opencl_pbkdf2_hmac_md5_fmt_plug.c
@@ -62,7 +62,7 @@ static int split_events[] = { 2, -1, -1 };
 static cl_kernel pbkdf2_init, pbkdf2_loop, pbkdf2_final;
 
 //This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 /* ------- Helper functions ------- */
