+++ src/opencl_pgpwde_fmt_plug.c
@@ -75,7 +75,7 @@ static struct fmt_main *self;
 size_t insize, outsize, settingsize, cracked_size;
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char *warn[] = {
