+++ src/opencl_rakp_fmt_plug.c
@@ -80,7 +80,7 @@ static int partial_output;
 static struct fmt_main *self;
 
 //This file contains auto-tuning routine(s). Have to included after other definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static struct fmt_tests tests[] = {
