+++ src/opencl_rawsha512_gpl_fmt_plug.c
@@ -97,7 +97,7 @@ static void release_kernel();
 static void release_mask_buffers(void);
 
 //This file contains auto-tuning routine(s). It has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 /* ------- Helper functions ------- */
