+++ src/opencl_strip_fmt_plug.c
@@ -98,7 +98,7 @@ static size_t insize, outsize, settingsize, cracked_size;
 #define SEED			256
 
 // This file contains auto-tuning routine(s). Has to be included after formats definitions.
-#include "opencl-autotune.h"
+#include "opencl_autotune.h"
 #include "memdbg.h"
 
 static const char * warn[] = {
